refactor: optimize profile code with utility functions and updated logo
