+++ src/app/student/_pages/EditProfilePage.tsx
@@ -72,6 +72,7 @@ import useCollegeSearch from '@/app/student/hooks/useCollegeSearch';
 import { toast } from '@/components/ui/use-toast';
 import { ProjectModal } from '@/app/student/profile/ProfileStep2';
 import { WorkExperienceModal, WorkExperienceCard } from '@/app/student/profile/WorkExperienceComponents';
+import { normalizeText, normalizeEmail, normalizeOptionalUrl, normalizeGithubUrl, parseScore, formatMonthYearToDate, toMonthNumber, toMonthName, mapWorkExperiencesToPayload } from '@/app/student/_utils/profileNormalizers';
 
 type TabType = 'basic-info' | 'skills-projects' | 'education' | 'career-goals';
 type EditingCard = 'personal-info' | 'skills' | 'projects' | 'academic-info' | 'academic-performance' | 'work-experience' | 'competitive-profiles' | 'career-goals' | null;
@@ -275,43 +276,7 @@ export const EditProfilePage: React.FC = () => {
     }));
   };
 
-  const formatMonthYearToDate = (
-    value?: { day?: string; month: string; year: string } | string
-  ) => {
-    if (!value) {
-      return undefined;
-    }
-
-    if (typeof value === 'string') {
-      const trimmed = value.trim();
-      return trimmed || undefined;
-    }
-
-    if (!value.year) {
-      return undefined;
-    }
-
-    const normalizedMonth = String(value.month || '').trim();
-    const numericMonth = Number(normalizedMonth);
-    let month = '01';
-
-    if (Number.isFinite(numericMonth) && numericMonth >= 1 && numericMonth <= 12) {
-      month = String(Math.trunc(numericMonth)).padStart(2, '0');
-    } else {
-      const monthIndex = MONTHS.findIndex(
-        (monthName) => monthName.toLowerCase() === normalizedMonth.toLowerCase()
-      );
-      month = monthIndex >= 0 ? String(monthIndex + 1).padStart(2, '0') : '01';
-    }
-
-    const numericDay = Number(String(value.day || '').trim());
-    const day =
-      Number.isFinite(numericDay) && numericDay >= 1 && numericDay <= 31
-        ? String(Math.trunc(numericDay)).padStart(2, '0')
-        : '01';
-
-    return `${value.year}-${month}-${day}`;
-  };
+  // formatMonthYearToDate moved to _utils/profileNormalizers.ts
 
   const formatWorkExperienceDate = (value?: { month: string; year: string } | string) => {
     if (!value) {
@@ -333,11 +298,7 @@ export const EditProfilePage: React.FC = () => {
     return `${value.month} ${value.year}`;
   };
 
-  const toMonthNumber = (month?: string) => {
-    if (!month) return undefined;
-    const monthIndex = MONTHS.indexOf(month);
-    return monthIndex >= 0 ? monthIndex + 1 : undefined;
-  };
+  // toMonthNumber, toMonthName moved to _utils/profileNormalizers.ts
 
   const buildLearnerProfilePayload = (
     step1Data: Step1Type,
@@ -345,44 +306,6 @@ export const EditProfilePage: React.FC = () => {
     step3Data?: Step3Type,
     step4Data?: Step4Type
   ) => {
-    const normalizeText = (value?: any) => {
-      if (value === null || value === undefined || value === '') return null;
-      const stringValue = String(value).trim();
-      return stringValue ? stringValue : null;
-    };
-
-    const normalizeEmail = (value?: any) => {
-      if (value === null || value === undefined || value === '') return null;
-      const stringValue = String(value).trim();
-      if (!stringValue) return null;
-      const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
-      return emailRegex.test(stringValue) ? stringValue : null;
-    };
-
-    const normalizeOptionalUrl = (value?: any) => {
-      if (value === null || value === undefined || value === '') return null;
-      const stringValue = String(value).trim();
-      if (!stringValue) return null;
-
-      const valueWithProtocol = /^[a-zA-Z][a-zA-Z\d+\-.]*:\/\//.test(stringValue)
-        ? stringValue
-        : `https://${stringValue}`;
-
-      try {
-        return new URL(valueWithProtocol).toString();
-      } catch {
-        return null;
-      }
-    };
-
-    const normalizeGithubUrl = (value?: any) => {
-      const normalizedUrl = normalizeOptionalUrl(value);
-      if (!normalizedUrl) return null;
-
-      const host = new URL(normalizedUrl).hostname.toLowerCase();
-      return host === 'github.com' || host === 'www.github.com' ? normalizedUrl : null;
-    };
-
     const technicalSkills = Array.from(
       new Set([...(step2Data?.autoDetectedSkills || []), ...(step2Data?.additionalSkills || [])].filter(Boolean))
     );
@@ -402,16 +325,7 @@ export const EditProfilePage: React.FC = () => {
     const academicPerformance = step3Data?.academicPerformance;
     const collegeScore = academicPerformance?.percentage;
 
-    const workExperiences = (step3Data?.workExperiences || []).map((experience) => ({
-      title: experience.role,
-      company: experience.companyName,
-      startDate: formatMonthYearToDate(experience.startDate),
-      isCurrentlyWorking: experience.isCurrentlyWorking,
-      endDate: experience.isCurrentlyWorking
-      ? null
-      : formatMonthYearToDate(experience.endDate),
-      description: experience.responsibilities,
-    }));
+    const workExperiences = mapWorkExperiencesToPayload(step3Data?.workExperiences || []);
 
     const leetcodeUsername = step3Data?.competitiveProfiles?.find((item) => item.platform === 'LeetCode')?.username;
     const codechefUsername = step3Data?.competitiveProfiles?.find((item) => item.platform === 'CodeChef')?.username;
@@ -527,31 +441,6 @@ export const EditProfilePage: React.FC = () => {
     return true;
   };
 
-  const toMonthName = (value?: string | number | null) => {
-    if (value === null || value === undefined) return '';
-    const monthNumber = Number(value);
-    const monthNames = [
-      'January',
-      'February',
-      'March',
-      'April',
-      'May',
-      'June',
-      'July',
-      'August',
-      'September',
-      'October',
-      'November',
-      'December',
-    ];
-    if (Number.isFinite(monthNumber) && monthNumber >= 1 && monthNumber <= 12) {
-      return monthNames[monthNumber - 1];
-    }
-    const normalized = String(value).trim().toLowerCase();
-    const mapped = monthNames.find((month) => month.toLowerCase() === normalized);
-    return mapped || '';
-  };
-
   useEffect(() => {
     if (!learnerProfile || hydratedProfileIdRef.current === learnerProfile.id) {
       return;
@@ -589,12 +478,6 @@ export const EditProfilePage: React.FC = () => {
       })),
       autoDetectedSkills: learnerProfile.technicalSkills || [],
       additionalSkills: [],
-    };
-
-    const parseScore = (value: string | number | null) => {
-      if (value === null || value === undefined || value === '') return undefined;
-      const parsed = Number(value);
-      return Number.isFinite(parsed) ? parsed : undefined;
     };
 
     const buildCompetitiveProfile = (
@@ -834,7 +717,7 @@ export const EditProfilePage: React.FC = () => {
   // FIX: read fresh from onboardingData at call time, not closed-over step3
   const currentStep3 = onboardingData?.step3;
 
-  if (step1 && currentStep3) {
+  if (step1) {
     const step1Edits = editedData.step1 || {};
     const shouldApplyCustomDegree =
       selectedDegreeValue === 'Other' &&
@@ -877,10 +760,14 @@ export const EditProfilePage: React.FC = () => {
 
     // FIX: build updatedStep3 from currentStep3 (fresh), not stale step3 closure
     const updatedStep3: Step3Type = {
-      ...currentStep3,
+      ...(currentStep3 || {
+        workExperiences: [],
+        competitiveProfiles: [],
+        hasInternshipExperience: false,
+      }),
       ...(editedData.step3 ? editedData.step3 : {}),
       // Always preserve fresh work experiences — never let editedData overwrite them
-          workExperiences: hasInternship ? currentStep3.workExperiences : [],
+          workExperiences: hasInternship ? (currentStep3?.workExperiences ?? []) : [],
       hasInternshipExperience: hasInternship,
     };
 
@@ -2057,7 +1944,7 @@ export const EditProfilePage: React.FC = () => {
                     <GraduationCap className="w-5 h-5 text-primary" />
                     <h3 className="text-base font-semibold tracking-wide">Academic Information</h3>
                   </div>
-                  {editingCard !== 'academic-info' && (step1?.collegeName || step1?.customCollege || step3?.academicPerformance) && (
+                  {editingCard !== 'academic-info' && step1 && (
                     <Button
                       variant="ghost"
                       size="sm"
@@ -2487,8 +2374,7 @@ export const EditProfilePage: React.FC = () => {
                         )}
                         
                         {/* Academic Performance Section */}
-                        {step3?.academicPerformance && (
-                          <>
+                        <>
                             <div className="grid gap-4 md:grid-cols-2">
                               <div className="space-y-2">
                                 <Label htmlFor="collegeScore" className="font-medium">College Marks *</Label>
@@ -2648,7 +2534,6 @@ export const EditProfilePage: React.FC = () => {
                               </div>
                             </div>
                           </>
-                        )}
                         
                         {/* Save/Cancel Buttons */}
                         <div className="flex justify-end gap-2 pt-4">
